fix spaces in column name upsert
