+++ data/sample_3.csv
@@ -1,0 +1,5 @@
+col a,col b,col_c
+1,updated value,E
+3,test,Z
+5,new value,F
+6,also new,H
